New Look
Updated the look of the website to not look like it was made in the early 2000's. Gave it a minimalistic look and fixed the header and footer. Fixed the gallery issue that I had, which made the images appear right next to the image gallery, distorting the page in the process. Now, I will need to just input the information in the wiki and then begin the process of changing the entirety of the other two pages.